basic inventory add command

+++ src/db/sql/setup/schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS users (
     id TEXT PRIMARY KEY,
-    points BIGINT,
+    points INT,
     duplicates INT,
     achievements SMALLINT
 );
